updated GravityPDF Extended plugin on server

+++ plugins/gravity-forms-pdf-extended/src/model/Model_PDF.php
@@ -1046,7 +1046,7 @@ class Model_PDF extends Helper_Abstract_Model {
 				}
 
 				/* Check if the file is too old and delete file / directory */
-				if ( filemtime( $file ) < time() - $max_file_age ) {
+				if ( file_exists( $file ) && filemtime( $file ) < time() - $max_file_age ) {
 
 					if ( $directory ) {
 						$this->misc->rmdir( substr( $file, 0, -1 ) );
@@ -1210,7 +1210,10 @@ class Model_PDF extends Helper_Abstract_Model {
 	 */
 	public function add_unregistered_fonts_to_mPDF( $fonts ) {
 
-		foreach ( glob( $this->data->template_font_location . '*.{otf,ttf,OTF,TTF}', GLOB_BRACE ) as $font ) {
+		$user_fonts = glob( $this->data->template_font_location . '*.{otf,ttf,OTF,TTF}', GLOB_BRACE );
+		$user_fonts = ( is_array( $user_fonts ) ) ? $user_fonts : array();
+
+		foreach ( $user_fonts as $font ) {
 
 			/* Get font shortname */
 			$font_name  = basename( $font );
